+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   `LivingStyleGuide.add_filter :name do |array_of_arguments, options_hash, block_string|`
 * `@font-example` works with multiple fonts separated by comma
 * Commas and quotation marks can be escaped in arguments/options (\, \" \')
+* Added `@style` commands to [customize the style guide’s own style](https://github.com/livingstyleguide/livingstyleguide/issues/126)
 
 
 ## 2.0.0.alpha.4